Solution of "Todos List" task
