+++ src/App.tsx
@@ -1,20 +1,17 @@
 import React from 'react';
 import './App.scss';
-
 import todosFromServer from './api/todos';
 import usersFromServer from './api/users';
-
-import { User } from './types/User';
-import { Todo } from './types/Todo';
+import { User, Todo } from './react-app-env';
+import { TodoList } from './components/TodoList';
 
 function getUser(userId: number): User | null {
   const foundUser = usersFromServer.find(user => user.id === userId);
 
-  // if there is no user with a given userId
   return foundUser || null;
 }
 
-export const todos: Todo[] = todosFromServer.map(todo => ({
+export const todos: Todo[] = todosFromServer.map((todo) => ({
   ...todo,
   user: getUser(todo.userId),
 }));
@@ -22,39 +19,6 @@ export const todos: Todo[] = todosFromServer.map(todo => ({
 export const App: React.FC = () => (
   <div className="App">
     <h1 className="App__title">Static list of todos</h1>
-
-    <section className="TodoList">
-      <article className="TodoInfo TodoInfo--completed">
-        <h2 className="TodoInfo__title">HTML</h2>
-
-        <a className="UserInfo" href="mailto:Sincere@april.biz">
-          Leanne Graham
-        </a>
-      </article>
-
-      <article className="TodoInfo TodoInfo--completed">
-        <h2 className="TodoInfo__title">CSS</h2>
-
-        <a className="UserInfo" href="mailto:Sincere@april.biz">
-          Leanne Graham
-        </a>
-      </article>
-
-      <article className="TodoInfo TodoInfo--completed">
-        <h2 className="TodoInfo__title">JS</h2>
-
-        <a className="UserInfo" href="mailto:Shanna@melissa.tv">
-          Ervin Howell
-        </a>
-      </article>
-
-      <article className="TodoInfo">
-        <h2 className="TodoInfo__title">React</h2>
-
-        <a className="UserInfo" href="mailto:Nathan@yesenia.net">
-          Clementine Bauch
-        </a>
-      </article>
-    </section>
+    <TodoList todos={todos} />
   </div>
 );
