+++ src/components/TodoList/TodoList.tsx
@@ -1,4 +1,15 @@
-// Add the required types and props
-export const TodoList = () => (
-  <>TodoList markup</>
+import React from 'react';
+import { TodoInfo } from '../TodoInfo';
+import { Todo } from '../../react-app-env';
+
+type Props = {
+  todos: Todo[],
+};
+
+export const TodoList: React.FC<Props> = ({ todos }) => (
+  <section className="TodoList">
+    {todos.map((todo) => (
+      <TodoInfo todo={todo} key={todo.id} />
+    ))}
+  </section>
 );
