--- src/types/Todo.ts
@@ -1,9 +1,0 @@
-import { User } from './User';
-
-export interface Todo {
-  id: number,
-  userId: number,
-  title: string,
-  completed: boolean,
-  user: User | null,
-}
