--- src/types/User.ts
@@ -1,5 +1,0 @@
-export interface User {
-  id: number,
-  name: string,
-  email: string,
-}
